code_wars_java/src/test/java/com/inertia/integers/BinaryArrayToNumberTest.java
package com.inertia.integers;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import static org.junit.Assert.*;

public class BinaryArrayToNumberTest {

    @org.junit.Test
    public void test1() throws Exception {
        assertEquals(1, BinaryArrayToNumber.ConvertBinaryArrayToInt(new ArrayList<>(Arrays.asList(0,0,0,1))));
    }

    @org.junit.Test
    public void test2() throws Exception {
        assertEquals(15, BinaryArrayToNumber.ConvertBinaryArrayToInt(new ArrayList<>(Arrays.asList(1,1,1,1))));
    }
    @org.junit.Test
    public void test3() throws Exception {
        assertEquals(6, BinaryArrayToNumber.ConvertBinaryArrayToInt(new ArrayList<>(Arrays.asList(0,1,1,0))));

    }
    @org.junit.Test
    public void test4() throws Exception {
        assertEquals(9, BinaryArrayToNumber.ConvertBinaryArrayToInt(new ArrayList<>(Arrays.asList(1,0,0,1))));
    }

}
